services/sitrep-service/requirements.txt
# services/sitrep-service/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
aio-pika>=9.3.0
openai>=1.8.0
httpx>=0.26.0
prometheus-client>=0.19.0
python-json-logger>=2.0.7
python-jose[cryptography]>=3.3.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Shared libraries (installed in Dockerfile)
news-intelligence-common
news-mcp-common
